requirements.txt
openai==1.17.0
streamlit==1.33.0
pypdf2==3.0.1
python-dotenv